LPS-64935 Wrong package name
