+++ modules/apps/web-experience/asset/asset-display-web/src/main/resources/META-INF/resources/init.jsp
@@ -22,7 +22,7 @@ taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %><%@
 taglib uri="http://liferay.com/tld/util" prefix="liferay-util" %>
 
-<%@ page import="com.liferay.asset.browser.web.display.context.AssetDisplayDisplayContext" %>
+<%@ page import="com.liferay.asset.display.web.display.context.AssetDisplayDisplayContext" %>
 
 <liferay-frontend:defineObjects />
